requirements.txt
# Core AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Configuration management
PyYAML>=6.0
python-dotenv>=1.0.0

# CLI interface
click>=8.1.0
rich>=13.0.0

# Data processing
pandas>=2.0.0

# Report generation
jinja2>=3.1.0
matplotlib>=3.7.0

# Date/Time handling
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0  # AWS mocking for tests

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
